add unit tests for the entity validators
