+++ packages/core/strapi/lib/services/entity-validator/__tests__/enumeration-validators.test.js
@@ -1,0 +1,41 @@
+'use strict';
+
+const strapiUtils = require('@strapi/utils');
+const { YupValidationError } = require('@strapi/utils/lib/errors');
+const entityValidator = require('../validators');
+
+describe('Enumeration validator', () => {
+  describe('oneOf', () => {
+    test('it fails the validation if the value is not part of the allowed values', async () => {
+      expect.assertions(1);
+
+      const validator = strapiUtils.validateYupSchema(
+        entityValidator.enumeration(
+          {
+            attr: { type: 'enum', enum: ['strapi', 'headless'] },
+          },
+          { isDraft: false }
+        )
+      );
+
+      try {
+        await validator('invalid-vlue');
+      } catch (err) {
+        expect(err).toBeInstanceOf(YupValidationError);
+      }
+    });
+
+    test('it validates the value if it is part of the allowed values', async () => {
+      const validator = strapiUtils.validateYupSchema(
+        entityValidator.enumeration(
+          {
+            attr: { type: 'enumeration', enum: ['strapi', 'headless'] },
+          },
+          { isDraft: false }
+        )
+      );
+
+      expect(await validator('strapi')).toBe('strapi');
+    });
+  });
+});
